fix : indicator 네비게이션바 아이템 인디케이터 투명색으로
