+++ app/src/main/java/com/echoist/linkedout/data/BottomNavItem.kt
@@ -1,0 +1,37 @@
+package com.echoist.linkedout.data
+
+import com.echoist.linkedout.R
+
+sealed class BottomNavItem(
+    val title: String,
+    val icon: Int,
+    val screenRoute: String
+) {
+    data object Home :
+        BottomNavItem(
+            "HOME",
+            R.drawable.bottom_nav_1,
+            "HOME" // nav Route 전부 재지정 필요
+        )
+
+    data object MyLog :
+        BottomNavItem(
+            "MYLOG",
+            R.drawable.bottom_nav_2,
+            "MYLOG"
+        )
+
+    data object Community :
+        BottomNavItem(
+            "COMMUNITY",
+            R.drawable.bottom_nav_3,
+            "COMMUNITY"
+        )
+
+    data object Settings :
+        BottomNavItem(
+            "SETTINGS",
+            R.drawable.bottom_nav_4,
+            "SETTINGS"
+        )
+}
